+++ package.json
@@ -39,9 +39,11 @@
     "@nestjs/common": "^6.0.0"
   },
   "dependencies": {
+    "@nestjs/axios": "^1.0.0",
     "@nestjs/class-validator": "^0.13.4",
     "@types/lodash": "^4.14.186",
     "@types/superagent": "^4.1.15",
+    "axios": "^1.1.3",
     "lodash": "^4.17.21",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.5.7"
@@ -79,6 +81,11 @@
       "^.+\\.(t|j)s$": "ts-jest"
     },
     "coverageDirectory": "../coverage",
-    "testEnvironment": "node"
+    "testEnvironment": "node",
+    "moduleDirectories": [
+      "node_modules",
+      "src",
+      "./"
+    ]
   }
 }
