+++ src/core/module/imports.ts
@@ -1,0 +1,5 @@
+import { HttpModule } from '@nestjs/axios';
+
+export default function() {
+  return [HttpModule];
+}
