+++ src/tests/fake/module-options.ts
@@ -1,0 +1,8 @@
+import { ZarinpalModuleOptions } from 'src/core/schema/interfaces';
+
+export const fakeModuleOptions: ZarinpalModuleOptions = {
+  callBackUrl: 'https://google.com',
+  clientName: 'Ziraxi',
+  merchantID: 2342346634234566,
+  sandboxMode: true,
+};
