--- src/tests/zarinpal-services.spec.ts
@@ -1,29 +1,0 @@
-import { ModuleMocker, MockFunctionMetadata } from 'jest-mock';
-import { Test } from '@nestjs/testing';
-const moduleMocker = new ModuleMocker(global);
-
-describe('CatsController', () => {
-  let controller: CatsController;
-
-  beforeEach(async () => {
-    const moduleRef = await Test.createTestingModule({
-      controllers: [CatsController],
-    })
-      .useMocker(token => {
-        const results = ['test1', 'test2'];
-        if (token === CatsService) {
-          return { findAll: jest.fn().mockResolvedValue(results) };
-        }
-        if (typeof token === 'function') {
-          const mockMetadata = moduleMocker.getMetadata(
-            token,
-          ) as MockFunctionMetadata<any, any>;
-          const Mock = moduleMocker.generateFromMetadata(mockMetadata);
-          return new Mock();
-        }
-      })
-      .compile();
-
-    controller = moduleRef.get(CatsController);
-  });
-});
